refactor: split into separate modules and refactor directory controller logic
